Move cmToReadable into formatting_logic.cpp with a covering test
helpers.cpp was almost entirely preferences I/O with three small tail
functions hitching a ride. cmToReadable was the only one truly pure;
calcMovingAvg and getFocusRadius both mutate globals, so extracting
them would add more bookkeeping than it removes. Pull cmToReadable
into its own pure module so the cm/m boundary and the precision
parameter get unit coverage.
