+++ Firmware/include/helpers.h
@@ -22,8 +22,6 @@ void savePrefs(bool force = false, uint8_t dirtyMask = PREFS_DIRTY_ALL);
 void flushPrefsIfDirty();
 void performFactoryReset();
 
-void cmToReadable(int cm, int places, char *buffer, size_t bufferSize);
-
 int calcMovingAvg(int sensorVal);
 
 int_fast16_t getFocusRadius();
